+++ src/Http/Controllers/RestfulController.php
@@ -2,12 +2,10 @@
 
 namespace Rougin\Weasley\Http\Controllers;
 
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
 use Rougin\Weasley\Validators\AbstractValidator;
 
 /**