+++ tests/Fixture/Templates/TestMiddleware.php
@@ -15,8 +15,8 @@ use Psr\Http\Message\ServerRequestInterface;
 class TestMiddleware implements MiddlewareInterface
 {
     /**
-     * Process an incoming server request and return a response, optionally delegating
-     * to the next middleware component to create the response.
+     * Process an incoming server request and return a response, optionally
+     * delegating to the next middleware component to create the response.
      *
      * @param  \Psr\Http\Message\ServerRequestInterface         $request
      * @param  \Interop\Http\ServerMiddleware\DelegateInterface $delegate