+++ tests/Http/MiddlewareTest.php
@@ -35,8 +35,8 @@ class MiddlewareTest extends \PHPUnit_Framework_TestCase
         array_push($stack, 'Rougin\Weasley\Http\Middleware\Cors');
         array_push($stack, 'Rougin\Weasley\Http\Middleware\FormMethodSpoofing');
         array_push($stack, 'Rougin\Weasley\Http\Middleware\TransformRequest');
-        array_push($stack, 'Rougin\Weasley\Http\Middleware\EmptyStringsToNull');
-        array_push($stack, 'Rougin\Weasley\Http\Middleware\TrimStrings');
+        array_push($stack, 'Rougin\Weasley\Http\Middleware\EmptyStringToNull');
+        array_push($stack, 'Rougin\Weasley\Http\Middleware\TrimString');
         array_push($stack, 'Rougin\Slytherin\Middleware\FinalResponse');
 
         $this->response = $middleware($request, $response, $stack);